Add ability to have cursor start on a new line
In order for the right prompt to work the new line must start with at
least one character. A space is used for this purpose. This looks a bit
strange so added option to include the default right arrow glyph. Lastly,
added option to remove the right arrow glyph on the first (top) line.