Use repo root when looking for untracked files
Currently, if you add a new file to a directory, then cd to a sibling directory, the untracked glyph will not appear. This is inconsistent with the behavior of the other git colors and flags, e.g. if you modify a file and cd into a sibling directory, the prompt will still be red and the dirty working directory glyph will still appear.

This PR modifies the query used to find untracked files so it searches from the repo root instead of the current directory.