Rework `__bobthefish_prompt_golang` to be more like "rubies" prompt
First checks to see if `asdf` is installed. If it is, checks to see if either of
the plugins ( `golang` or `go-sdk` ) that install a version of Go have an
installed version. If they do, use that version.

If `asdf` isn't installed, check the `go.mod` file. If in a folder that has a
`go.mod` or a parent folder has a `go.mod`, check that file for the Go version.

Lastly, check if there's a `go` executable on the current path. If there isn't
set a flag so that the go version will be printed out in red rather than blue.